salary page fix
